% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assets.R
\name{get_assets}
\alias{get_assets}
\title{Calculate Total Assets by Account Description}
\usage{
get_assets(ledger_file = TRUE)
}
\arguments{
\item{ledger_file}{Path to the CSV ledger file}
}
\value{
A data frame containing account descriptions and their total asset
values
}
\description{
This function reads a ledger file and calculates the sum of assets for each
account description, considering both debit and credit entries that start with
'1'.
}
\examples{
\dontrun{
get_assets(ledger_file = my_ledger_directory)
}
}
