% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger.R
\name{get_ledger}
\alias{get_ledger}
\title{Get Last Ledger Version}
\usage{
get_ledger(ledger_file = NULL, import_csv = FALSE)
}
\arguments{
\item{ledger_file}{Path to the CSV ledger file}

\item{import_csv}{Logical indicating whether to import from CSV}
}
\value{
A tibble containing ledger data or an empty ledger structure if no file
is provided
}
\description{
Retrieves the most recent version of the ledger, either creating an empty one or
importing from a CSV file.
}
\keyword{internal}
