% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissaccounting-package.R
\docType{package}
\name{swissaccounting-package}
\alias{swissaccounting}
\alias{swissaccounting-package}
\title{swissaccounting: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Layal Christine Lettry \email{layal@cynkra.com}

}
\keyword{internal}
