% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_type.R
\name{try_ensure_type}
\alias{try_ensure_type}
\title{Try to ensure type stability for data frame columns}
\usage{
try_ensure_type(.data, ..., .default = NULL, .call = rlang::caller_env())
}
\arguments{
\item{.data}{A data frame whose columns should be cast to specified types.}

\item{...}{Named arguments specifying the desired type for each column.}

\item{.default}{Optional default type to apply to unspecified columns.}

\item{.call}{The call environment for error reporting.}
}
\value{
A tibble with columns cast to the specified types.
}
\description{
Internal function that attempts to cast columns to specified types.
This is the workhorse function called by \code{ensure_type()}.
}
\keyword{internal}
